svdb/integrations/cloud_adapters/azure_adapter.py
"""
Azure Blob Storage Adapter for SVDB
"""
import io
import os
import logging
from typing import List, Dict, Any, BinaryIO, Optional, Tuple

from . import CloudStorageAdapter

logger = logging.getLogger("svdb.integrations.azure")

class AzureBlobAdapter(CloudStorageAdapter):
    """Azure Blob Storage Adapter"""
    
    def __init__(
        self, 
        account_name: str,
        container_name: str,
        account_key: Optional[str] = None,
        connection_string: Optional[str] = None
    ):
        """
        Initialize Azure Blob Storage adapter
        
        Args:
            account_name: Azure Storage account name
            container_name: Container name
            account_key: Account key (optional if connection_string is provided)
            connection_string: Connection string (optional if account_name and account_key are provided)
        """
        self.account_name = account_name
        self.container_name = container_name
        self.account_key = account_key
        self.connection_string = connection_string
        self.blob_service_client = None
        self.container_client = None
        
        # For storing import results to avoid duplicate imports
        self.import_cache = {}
    
    def connect(self) -> bool:
        """Connect to Azure Blob Storage"""
        try:
            # Required package
            from azure.storage.blob import BlobServiceClient
            
            # Create clients
            if self.connection_string:
                self.blob_service_client = BlobServiceClient.from_connection_string(self.connection_string)
            else:
                # For token-based auth or using account key
                account_url = f"https://{self.account_name}.blob.core.windows.net"
                self.blob_service_client = BlobServiceClient(account_url, credential=self.account_key)
            
            # Get container client
            self.container_client = self.blob_service_client.get_container_client(self.container_name)
            
            # Verify connection by fetching container properties
            self.container_client.get_container_properties()
            return True
        
        except ImportError:
            logger.error("azure-storage-blob package is required for Azure adapter")
            return False
        
        except Exception as e:
            logger.error(f"Failed to connect to Azure Blob Storage: {e}")
            return False
    
    def list_files(self, prefix: str = "") -> List[Dict[str, Any]]:
        """List files in Azure Blob Storage container with optional prefix"""
        # Implementation needed
        logger.warning("Azure adapter list_files method not fully implemented")
        return []
    
    def get_file(self, file_key: str) -> BinaryIO:
        """Get a file from Azure Blob Storage as a file-like object"""
        # Implementation needed
        logger.warning("Azure adapter get_file method not fully implemented")
        return io.BytesIO(b"")
    
    def get_file_metadata(self, file_key: str) -> Dict[str, Any]:
        """Get metadata for a file in Azure Blob Storage"""
        # Implementation needed
        logger.warning("Azure adapter get_file_metadata method not fully implemented")
        return {
            'key': file_key,
            'size': 0,
            'last_modified': None,
        }
    
    def import_file(self, file_key: str, svdb_path: str) -> Tuple[str, Optional[str]]:
        """Import a file from Azure Blob Storage to SVDB"""
        # Implementation needed
        logger.warning("Azure adapter import_file method not fully implemented")
        return ("mock_hash", None)
    
    def import_files(self, file_keys: List[str], svdb_path: str) -> Dict[str, Tuple[str, Optional[str]]]:
        """Import multiple files from Azure Blob Storage to SVDB"""
        # Implementation needed
        logger.warning("Azure adapter import_files method not fully implemented")
        return {file_key: ("mock_hash", None) for file_key in file_keys}

# Convenience function to create adapter from environment variables
def create_azure_adapter_from_env() -> AzureBlobAdapter:
    """
    Create Azure adapter from environment variables
    
    Required environment variables:
    - AZURE_STORAGE_ACCOUNT: Azure Storage account name
    - AZURE_STORAGE_CONTAINER: Container name
    
    Optional environment variables (at least one of these is required):
    - AZURE_STORAGE_KEY: Account key
    - AZURE_STORAGE_CONNECTION_STRING: Connection string
    
    Returns:
        Configured AzureBlobAdapter instance
    """
    account_name = os.getenv('AZURE_STORAGE_ACCOUNT')
    container_name = os.getenv('AZURE_STORAGE_CONTAINER')
    
    if not account_name or not container_name:
        raise ValueError("AZURE_STORAGE_ACCOUNT and AZURE_STORAGE_CONTAINER environment variables are required")
    
    return AzureBlobAdapter(
        account_name=account_name,
        container_name=container_name,
        account_key=os.getenv('AZURE_STORAGE_KEY'),
        connection_string=os.getenv('AZURE_STORAGE_CONNECTION_STRING')
    ) 