svdb/integrations/cloud_adapters/gcp_adapter.py
"""
Google Cloud Storage Adapter for SVDB
"""
import io
import os
import logging
from typing import List, Dict, Any, BinaryIO, Optional, Tuple

from . import CloudStorageAdapter

logger = logging.getLogger("svdb.integrations.gcp")

class GCPStorageAdapter(CloudStorageAdapter):
    """Google Cloud Storage Adapter"""
    
    def __init__(
        self, 
        bucket_name: str,
        credentials_file: Optional[str] = None,
        project_id: Optional[str] = None
    ):
        """
        Initialize Google Cloud Storage adapter
        
        Args:
            bucket_name: GCS bucket name
            credentials_file: Path to GCP credentials JSON file (optional)
            project_id: GCP project ID (optional)
        """
        self.bucket_name = bucket_name
        self.credentials_file = credentials_file
        self.project_id = project_id
        self.storage_client = None
        self.bucket = None
        
        # For storing import results to avoid duplicate imports
        self.import_cache = {}
        
        # Set credentials environment variable if provided
        if credentials_file:
            os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = credentials_file
    
    def connect(self) -> bool:
        """Connect to Google Cloud Storage"""
        try:
            # Required package
            from google.cloud import storage
            
            # Create storage client
            self.storage_client = storage.Client(project=self.project_id)
            
            # Get bucket
            self.bucket = self.storage_client.bucket(self.bucket_name)
            
            # Verify connection by getting bucket metadata
            self.bucket.reload()
            return True
        
        except ImportError:
            logger.error("google-cloud-storage package is required for GCP adapter")
            return False
        
        except Exception as e:
            logger.error(f"Failed to connect to Google Cloud Storage: {e}")
            return False
    
    def list_files(self, prefix: str = "") -> List[Dict[str, Any]]:
        """List files in GCS bucket with optional prefix"""
        # Implementation needed
        logger.warning("GCP adapter list_files method not fully implemented")
        return []
    
    def get_file(self, file_key: str) -> BinaryIO:
        """Get a file from GCS as a file-like object"""
        # Implementation needed
        logger.warning("GCP adapter get_file method not fully implemented")
        return io.BytesIO(b"")
    
    def get_file_metadata(self, file_key: str) -> Dict[str, Any]:
        """Get metadata for a file in GCS"""
        # Implementation needed
        logger.warning("GCP adapter get_file_metadata method not fully implemented")
        return {
            'key': file_key,
            'size': 0,
            'last_modified': None,
        }
    
    def import_file(self, file_key: str, svdb_path: str) -> Tuple[str, Optional[str]]:
        """Import a file from GCS to SVDB"""
        # Implementation needed
        logger.warning("GCP adapter import_file method not fully implemented")
        return ("mock_hash", None)
    
    def import_files(self, file_keys: List[str], svdb_path: str) -> Dict[str, Tuple[str, Optional[str]]]:
        """Import multiple files from GCS to SVDB"""
        # Implementation needed
        logger.warning("GCP adapter import_files method not fully implemented")
        return {file_key: ("mock_hash", None) for file_key in file_keys}

# Convenience function to create adapter from environment variables
def create_gcp_adapter_from_env() -> GCPStorageAdapter:
    """
    Create GCP adapter from environment variables
    
    Required environment variables:
    - GCP_STORAGE_BUCKET: GCS bucket name
    
    Optional environment variables:
    - GOOGLE_APPLICATION_CREDENTIALS: Path to credentials JSON file
    - GCP_PROJECT_ID: GCP project ID
    
    Returns:
        Configured GCPStorageAdapter instance
    """
    bucket_name = os.getenv('GCP_STORAGE_BUCKET')
    
    if not bucket_name:
        raise ValueError("GCP_STORAGE_BUCKET environment variable is required")
    
    return GCPStorageAdapter(
        bucket_name=bucket_name,
        credentials_file=os.getenv('GOOGLE_APPLICATION_CREDENTIALS'),
        project_id=os.getenv('GCP_PROJECT_ID')
    ) 